% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day.R
\name{day_continue}
\alias{day_continue}
\title{Continue solving a puzzle from the Advent of Code}
\usage{
day_continue(date, path)
}
\arguments{
\item{date}{Day to fetch.}

\item{path}{Path to a file created by \code{\link[=day_start]{day_start()}}.}
}
\value{
Path to the appended file.
}
\description{
Given a file created by \code{\link[=day_start]{day_start()}}, append the 2nd part of a puzzle to the
end of said file.
}
\section{Authentication}{
This function is capable of using an authenticated session to fetch the 2nd
part of a puzzle after the user has answered the 1st part. For this to work,
the user must supply their session cookie as an environment variable called
\code{AOC_SESSION}.
}

\seealso{
\code{\link[=day_start]{day_start()}}
}
