% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day.R
\name{day_start}
\alias{day_start}
\title{Start solving a puzzle from the Advent of Code}
\usage{
day_start(date = Sys.Date(), path = "./")
}
\arguments{
\item{date}{Day to fetch.}

\item{path}{Where to create directory.}
}
\value{
Path to the created directory.
}
\description{
Create a directory with everything the user needs to start solving a puzzle
from the Advent of Code. Once the 1st part is done, \code{\link[=day_continue]{day_continue()}} appends
the 2nd part to the file created by this function.
}
\seealso{
\code{\link[=day_continue]{day_continue()}}
}
