% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{fetch_input}
\alias{fetch_input}
\title{Fetch the input of a puzzle from the Advent of Code}
\usage{
fetch_input(date = Sys.Date())
}
\arguments{
\item{date}{Day to fetch.}
}
\value{
The input of the puzzle.
}
\description{
Fetch the input of a puzzle from the Advent of Code given the day it was
published. This function automatically reads the input line by line and
writes it to the clipboard.
}
\section{Authentication}{
This function needs the user's session cookie in order to work because the
inputs vary from person to person. You must supply your session cookie as an
environment variable called \code{AOC_SESSION}
}

