% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzle.R
\name{fetch_puzzle}
\alias{fetch_puzzle}
\title{Fetch a puzzle from the Advent of Code}
\usage{
fetch_puzzle(part = 1, date = Sys.Date())
}
\arguments{
\item{part}{Part to fetch (either 1 or 2). See Details for more information.}

\item{date}{Day to fetch.}
}
\value{
Text of the puzzle.
}
\description{
Fetch (part of) a puzzle from the Advent of Code given the day it was
published. This function automatically transforms the text of the puzzle into
an R multi-line comment and writes it to the clipboard.
}
\section{Tooltips}{
Hidden tooltips (which only show up on the website when one hovers over the
text) are transformed into inline footnotes on the final output.
}

\section{Authentication}{
This function is capable of using an authenticated session to fetch the 2nd
part of a puzzle after the user has answered the 1st part. For this to work,
the user must supply their session cookie as an environment variable called
\code{AOC_SESSION}.
}

