% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{fetch_title}
\alias{fetch_title}
\title{Fetch the title of a puzzle from the Advent of Code}
\usage{
fetch_title(date = Sys.Date())
}
\arguments{
\item{date}{Day to fetch.}
}
\value{
A character string.
}
\description{
Fetch the title of a puzzle from the Advent of Code given the day it was
published.
}
