% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{title_fetch}
\alias{title_fetch}
\title{Fetch an Advent of Code puzzle's title}
\usage{
title_fetch(date = Sys.Date())
}
\arguments{
\item{date}{Day to fetch.}
}
\description{
Fetch the title of a puzzle from the Advent of Code given the day it was
published. This function automatically writes the title to the clipboard.
}
\seealso{
\code{\link[=day_start]{day_start()}}
}
