% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input_fetch}
\alias{input_fetch}
\title{Fetch an Advent of Code puzzle's input}
\usage{
input_fetch(date = Sys.Date())
}
\arguments{
\item{date}{Day to fetch.}
}
\description{
Fetch the input of a puzzle from the Advent of Code given the day it was
published. This function automatically reads the input line by line and
writes it to the clipboard.
}
\seealso{
\code{\link[=day_start]{day_start()}}
}
